fixed modals
fixed modals for assignments
