+++ Skolpatrullen/WebAPI/Controllers/FileController.cs
@@ -13,6 +13,7 @@ using WebApp.ViewModels;
 
 namespace WebAPI.Controllers
 {
+    [ApiController]
     public class FileController : APIController
     {
         public FileController(Context context, ILogger<UserController> logger) : base(context, logger)
@@ -73,6 +74,44 @@ namespace WebAPI.Controllers
             {
                 response.Success = false;
                 response.FailureMessage = "Hämtade inga filer";
+            }
+            return response;
+        }
+        [HttpPost]
+        [Route("[controller]/UploadCourseFile")]
+        public APIResponse UploadCourseFile(CourseFileBody body)
+        {
+            APIResponse response = new APIResponse();
+
+            User user = _context.Users.SingleOrDefault(u => u.Id == body.UserId);
+
+
+            if (user != null && body.File.Length > 0)
+            {
+                File file = new File();
+                CourseFile coursefile = new CourseFile();
+
+                file.Binary = body.File;
+                file.UploadDate = body.UploadDate;
+                file.ContentType = body.ContentType;
+                file.Type = FileTypes.CourseFile;
+                file.Name = body.Name;
+
+                _context.Files.Add(file);
+                _context.SaveChanges();
+
+                coursefile.CourseId = body.CourseId;
+                coursefile.FileId = file.Id;
+
+                _context.CourseFiles.Add(coursefile);
+                _context.SaveChanges();
+
+                response.Success = true;
+            }
+            else
+            {
+                response.FailureMessage = "Filen laddades inte upp";
+                response.Success = false;
             }
             return response;
         }
