+++ Skolpatrullen/WebApp/Controllers/AppController.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
 using System.Text;
+using System.Text.Json;
 using System.Threading.Tasks;
 using WebApp.ViewModels;
