+++ Skolpatrullen/WebApp/Views/Course/AdminCourseDetails.cshtml
@@ -16,7 +16,7 @@
         </div>
         <div class="list-group">
             @await Html.PartialAsync("_UploadCourseFiles", new WebApp.ViewModels.UploadCourseFileViewModel { CourseId = Model.Id })
-            <a class="list-group-item" href="@Url.Action("AddCourseAssignment", "course", new Database.Models.Assignment { CourseId = Model.Id })">Lägg till inlämning</a>
+            @await Html.PartialAsync("_AddCourseAssignment", new Database.Models.Assignment { CourseId = Model.Id })
         </div>
     </div>
 
@@ -27,5 +27,3 @@
         </div>
     </div>
 </div>
-
-
