+++ Skolpatrullen/WebApp/Views/Course/_AddCourseAssignment.cshtml
@@ -1,43 +1,39 @@
 ﻿@model Database.Models.Assignment
 
 
-<a asp-action="CourseAssignmentModal" data-toggle="modal" data-target="#CourseAssignmentModal" class="list-group-item">Lägg till inlämning</a>
+<button data-toggle="modal" data-target="#CourseAssignmentModal" class="list-group-item">Lägg till inlämning</button>
 
 
-<div class="modal fade" id="CourseAssignmentModal" tabindex="-1" role="dialog" aria-labelledby="wat" aria-hidden="true">
+<div class="modal fade" id="CourseAssignmentModal" tabindex="-1" role="dialog" aria-labelledby="CourseAssignment" aria-hidden="true">
     <div class="modal-dialog" role="document">
-        <div class="modal-content">
+        <form asp-action="AddCourseAssignment" class="modal-content">
             <div class="modal-header">
-                <h5 class="modal-title" id="exampleModalLabel">Ladda upp fil</h5>
+                <h5 class="modal-title" id="CourseAssignment">Lägg till inlämning</h5>
                 <button type="button" class="close" data-dismiss="modal" aria-label="Close">
                     <span aria-hidden="true">&times;</span>
                 </button>
             </div>
-            <form asp-action="AddCourseAssignment">
-                <div class=" modal-body">
-                    <label asp-for="Name" class="control-label">Namn</label>
-                    <input asp-for="Name" type="text" class="form-control" />
-                    <span asp-validation-for="Name" class="text-danger"></span>
-                </div>
-                <div class=" modal-body">
-                    <label asp-for="Description" class="control-label">Text</label>
-                    <textarea asp-for="Description" type="text" class="form-control" />
-                    <span asp-validation-for="Description" class="text-danger"></span>
-                </div>
-                <div class=" modal-body">
-                    <label asp-for="Deadline" class="control-label">Deadline för inlämning</label>
-                    <input asp-for="Deadline" type="date" class="form-control" />
-                    <span asp-validation-for="Deadline" class="text-danger"></span>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" data-dismiss="modal">Avbryt</button>
-                    <input type="submit" asp-action="AddCourseAssignment" value="Spara" class="btn btn-primary" />
-                </div>
-            </form>
-        </div>
+            <div class=" modal-body">
+                <label asp-for="Name" class="control-label">Namn</label>
+                <input asp-for="Name" type="text" class="form-control" />
+                <span asp-validation-for="Name" class="text-danger"></span>
+
+                <label asp-for="Description" class="control-label">Text</label>
+                <textarea asp-for="Description" type="text" class="form-control"></textarea>
+                <span asp-validation-for="Description" class="text-danger"></span>
+
+                <label asp-for="Deadline" class="control-label">Deadline för inlämning</label>
+                <input asp-for="Deadline" type="date" class="form-control" />
+                <span asp-validation-for="Deadline" class="text-danger"></span>
+            </div>
+            <div class="modal-footer">
+                <button type="button" class="btn btn-secondary" data-dismiss="modal">Avbryt</button>
+                <input type="submit" asp-action="AddCourseAssignment" asp-route-courseId="@Model.CourseId" value="Spara" class="btn btn-primary" />
+            </div>
+        </form>
     </div>
 </div>
 
 @section Scripts {
-    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
-}
+        @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
+    }
